backend/requirements.txt
django==5.1.7
djangorestframework
djangorestframework-simplejwt
django-cors-headers
mysqlclient
Pillow
python-decouple
